Move device_gl_error() to gl_helpers.c
