+++ c_src/device/bcm.c
@@ -292,24 +292,3 @@ void device_clear_color(float red,
 {
   glClearColor(red, green, blue, alpha);
 }
-
-char* device_gl_error()
-{
-  GLenum err  = glGetError();
-  switch (err)
-  {
-  case GL_NO_ERROR: return NULL;
-  case GL_INVALID_ENUM: return "GL_INVALID_ENUM";
-  case GL_INVALID_VALUE: return "GL_INVALID_VALUE";
-  case GL_INVALID_OPERATION: return "GL_INVALID_OPERATION";
-  case GL_OUT_OF_MEMORY: return "GL_OUT_OF_MEMORY";
-#ifdef GL_STACK_UNDERFLOW
-  case GL_STACK_UNDERFLOW: return "GL_STACK_UNDERFLOW";
-#endif
-#ifdef GL_STACK_OVERFLOW
-  case GL_STACK_OVERFLOW: return "GL_STACK_OVERFLOW";
-#endif
-  case GL_INVALID_FRAMEBUFFER_OPERATION: return "GL_INVALID_FRAMEBUFFER_OPERATION";
-  default: return "GL_OTHER";
-  }
-}
